* use stable_whisper makes more percise timestamp * add progressbar for translating

+++ subtitle.py
@@ -3,6 +3,7 @@ import os
 from pathlib import Path
 import whisper
 from whisper.utils import get_writer
+from stable_whisper import modify_model
 import torch
 import sys
 import contextlib
@@ -19,6 +20,10 @@ import re
 from moviepy.editor import *
 
 from deep_translator import GoogleTranslator
+from deep_translator import DeeplTranslator
+
+import progressbar
+#import config as c
 
 class Subtitle:
     def __init__(self):
@@ -29,7 +34,10 @@ class Subtitle:
         self.lang = "en"
         self.video_file = None
 
+        self.translate_engine = "google"
+
         self.whisper_model = whisper.load_model(self.model+"."+self.lang)
+        modify_model(self.whisper_model)
 
         self.recognise_result = None
 
@@ -42,6 +50,7 @@ class Subtitle:
             else:
                 r = self.whisper_model.transcribe(file, fp16=torch.cuda.is_available())
             self.recognise_result = r
+            print (r)
         except:
             self.recognise_result = None
 
@@ -49,14 +58,15 @@ class Subtitle:
         if self.recognise_result == None or self.video_file == None:
             return
 
-        srt_wr = get_writer("srt", os.path.dirname(self.video_file))
+        #srt_wr = get_writer("srt", os.path.dirname(self.video_file))
         filename = "{}.{}.srt".format(Path(os.path.basename(self.video_file)).stem, self.lang)
 
         if fname != None:
             filename = os.path.basename(fname)
 
         print("save subtitle to file {}".format(filename))
-        srt_wr(self.recognise_result, filename)
+        self.recognise_result.to_srt_vtt(filename, word_level=False)
+        #srt_wr(self.recognise_result, filename)
 
     def __lang_map(self, lang):
         if lang == None:
@@ -76,11 +86,28 @@ class Subtitle:
 
         lang = self.__lang_map(lang)
         src_lang = self.__lang_map(self.lang)
-        gt = GoogleTranslator(source = src_lang, target = lang)
+
+        if self.translate_engine == "google":
+            trans = GoogleTranslator(source = src_lang, target = lang)
+        elif self.translate_engine == "deepl":
+            trans = translated = DeeplTranslator(api_key="your_api_key", source=src_lang, target=lang, use_free_api=True)
+        else:
+            trans = GoogleTranslator(source = src_lang, target = lang)
+
+        num_lines = sum(1 for _ in open(file, "r", encoding="utf-8"))
+        count = 0
+        pb = progressbar.ProgressBar(maxval = num_lines).start()
 
         with open(file, "r", encoding="utf-8") as ori_file:
             for line in ori_file:
-                res = gt.translate(line) + "\n"
+                count = count + 1
+                pb.update(count)
+                l = line
+                find = re.search(r'[^0-9\-\>:, ]', l.strip())
+                if find != None:
+                    res = trans.translate(line) + "\n"
+                else:
+                    res = line
                 result += res
 
         return result
@@ -93,7 +120,11 @@ class Subtitle:
         else:
             return
 
+        modify_model(self.whisper_model)
         self.lang = lang
+
+    def set_translator_engine(self, engine):
+        self.translate_engine = engine
 
     def start_recognise(self):
         self.start_recognise_from_file(self.audio_tmpfile)
@@ -101,6 +132,14 @@ class Subtitle:
     def start_recognise_from_file(self, audio_file):
         print("recognise in ai model ...")
         self.__recognise(audio_file)
+
+    def __write_lang_str(self, srcfile, destfile, lang):
+        result = self.__translate(lang, srcfile)
+
+        print("create subtitle file {} ...".format(destfile))
+        f = open(destfile, 'w', encoding="utf-8")
+        f.write(result)
+        f.close()
 
     def store_to_srt(self, lang = None):
         if lang == None or lang == self.lang:
@@ -114,24 +153,34 @@ class Subtitle:
 
         print("create tmp subtitle file ...")
         self.srt_tmpfile = os.path.basename(NamedTemporaryFile().name) + ".srt"
-        self.__result_to_subtitle(self.srt_tmpfile)
-
-        result = self.__translate(lang, self.srt_tmpfile)
         if len(os.path.dirname(self.video_file)) == 0:
             filename = "{}.{}.srt".format(Path(os.path.basename(self.video_file)).stem, lang)
         else:
             filename = "{}/{}.{}.srt".format(os.path.dirname(self.video_file), Path(os.path.basename(self.video_file)).stem, lang)
-        print("create subtitle file {} ...".format(filename))
-        f = open(filename, 'w', encoding="utf-8")
-        f.write(result)
-        f.close()
+        self.__result_to_subtitle(self.srt_tmpfile)
+
+        self.__write_lang_str(self.srt_tmpfile, filename, lang)
+
+    def store_to_srt_from_srt(self, file, lang = None):
+        if file == None or not os.path.exists(file):
+            return
+        elif lang == "zh" or lang == "en" or lang == "jp":
+            print("start translate process...")
+        else:
+            return
+
+        if len(os.path.dirname(file)) == 0:
+            filename = "{}.{}.srt".format(Path(os.path.basename(file)).stem, lang)
+        else:
+            filename = "{}/{}.{}.srt".format(os.path.dirname(file), Path(os.path.basename(file)).stem, lang)
+        self.__write_lang_str(file, filename, lang)
 
     def close(self):
         print("remove tmp files....")
-        if os.path.exists(self.audio_tmpfile):
+        if self.audio_tmpfile != None and os.path.exists(self.audio_tmpfile):
             os.remove(self.audio_tmpfile)
 
-        if os.path.exists(self.srt_tmpfile):
+        if self.srt_tmpfile != None and os.path.exists(self.srt_tmpfile):
             os.remove(self.srt_tmpfile)
 
         self.recognise_result = None
